+++ Lucky Sevens/Assets/Scripts/Interfaces/IDamage.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IDamage
 {
-    void takeDamage(int dmg);
+    void takeDamage(float dmg);
 }
